Test/unit tests (#25)
* feat: Unit testing de los Fantasy Teams

Se agregaron los .py para las pruebas unitarias los metodos de los Fantasy Teams.
Se arreglaron las rutas de los modulos para que fucionaran las pruebas unitarias.
Se agrego un nuevo paso a realizar en el Readme.md para correr el servidor.

* feat: Unit testing de las ligas

Se agregaron los modulos de unit test para el repository.py y el router.py
Minimo cambio en router.py con una variable con el fin de evitar una colision

+++ backend/src/modules/fantasy_teams/models.py
@@ -1,10 +1,11 @@
 from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime, func, ForeignKey
-from ...config.database import Base
+from config.database import Base
 
 class FantasyTeam(Base):
     __tablename__ = "fantasy_teams"
 
-    id = Column(BigInteger, primary_key=True, index=True)
+    # Use Integer for primary key so SQLite will autoincrement correctly in tests
+    id = Column(Integer, primary_key=True, index=True)
     name = Column(String(128), nullable=False)
     image_url = Column(String(512), nullable=True)
     thumbnail_url = Column(String(512), nullable=True)
